Enforce code coherence: error handling, zero-means-zero, structural naming
Bounded storage violated QNTX LAW: <= 0 silently fell back to defaults
instead of treating 0 as literal zero. Now only negative values fall
back; 0 means "retain 0 attestations" as the principle demands.

Replaced fmt.Errorf with project errors package across cmd/, gopls, and
plugin code. The codebase defines github.com/teranos/QNTX/errors for
stack traces, context wrapping, and PII safety — using fmt.Errorf
bypasses all of it. Added ErrServiceNotInitialized sentinel to gopls
service to replace 6 identical fmt.Errorf("service not initialized")
calls.

Split glyph/handlers/canvas.go (593 lines) into canvas.go + subscriptions.go
to match the test file structure (canvas_test.go, subscriptions_test.go,
edge_cursor_test.go). Subscription compilation is a distinct domain from
CRUD routing.
