+++ ats/storage/bounded_store.go
@@ -53,14 +53,15 @@ func NewBoundedStoreWithConfig(db *sql.DB, logger *zap.SugaredLogger, config *Bo
 		config = DefaultBoundedStoreConfig()
 	}
 
-	// Validate config: zero or negative limits are invalid (use positive limits)
-	if config.ActorContextLimit <= 0 {
+	// Negative limits are invalid; default via am/defaults.go handles "omit = use default".
+	// Zero means zero (QNTX LAW): 0 limit = no attestations retained for that dimension.
+	if config.ActorContextLimit < 0 {
 		config.ActorContextLimit = DefaultActorContextLimit
 	}
-	if config.ActorContextsLimit <= 0 {
+	if config.ActorContextsLimit < 0 {
 		config.ActorContextsLimit = DefaultActorContextsLimit
 	}
-	if config.EntityActorsLimit <= 0 {
+	if config.EntityActorsLimit < 0 {
 		config.EntityActorsLimit = DefaultEntityActorsLimit
 	}
