+++ cmd/qntx/commands/database.go
@@ -2,10 +2,10 @@ package commands
 
 import (
 	"database/sql"
-	"fmt"
 
 	"github.com/teranos/QNTX/am"
 	"github.com/teranos/QNTX/db"
+	"github.com/teranos/QNTX/errors"
 	"github.com/teranos/QNTX/logger"
 )
 
@@ -16,7 +16,7 @@ func openDatabase(dbPath string) (*sql.DB, error) {
 	if dbPath == "" {
 		path, err := am.GetDatabasePath()
 		if err != nil {
-			return nil, fmt.Errorf("failed to get database path: %w", err)
+			return nil, errors.Wrap(err, "failed to get database path")
 		}
 		if path == "" {
 			dbPath = "qntx.db"
@@ -28,13 +28,13 @@ func openDatabase(dbPath string) (*sql.DB, error) {
 	// Open database with logger
 	database, err := db.Open(dbPath, logger.Logger)
 	if err != nil {
-		return nil, fmt.Errorf("failed to open database: %w", err)
+		return nil, errors.Wrapf(err, "failed to open database at %s", dbPath)
 	}
 
 	// Run migrations with logger
 	if err := db.Migrate(database, logger.Logger); err != nil {
 		database.Close()
-		return nil, fmt.Errorf("failed to run migrations: %w", err)
+		return nil, errors.Wrapf(err, "failed to run migrations on %s", dbPath)
 	}
 
 	return database, nil
